old-files/to-del/backend/requirements.txt
python-dotenv
langchain
langchain-google-genai
chromadb
PyPDF2
pytesseract
pdf2image
Pillow
sentence-transformers
python-dotenv
langchain
langchain-google-genai
chromadb
PyPDF2
pytesseract
pdf2image
Pillow
sentence-transformers
Flask
Flask-SocketIO
Flask-SQLAlchemy
Flask-Cors
eventlet
gunicorn
nest_asyncio